Clean up TOC links
Many of the TOC links linked to either the wrong section (due to numbering issues, etc) or no section at all (due to mismatches between the <a> tag and the markdown link). See for example any of the Julia links, which link to JavaScript sections.

This PR cleans up the naming system by replacing the numbering system with a standardized system of programming-language-section-name. It updates all of the TOC so that they now link to the correct sections.